feat: Create chart component
