+++ src/js/index.js
@@ -6,82 +6,6 @@ import { PopupMenuLinks } from "./PopupMenuLinks";
 import { MenuItemLinks } from "./MenuItemLinks";
 export { initFonts } from "./initFonts";
 
-import bb, { line } from "billboard.js";
-import "billboard.js/dist/billboard.css";
-
-export const lineChart = function () {
-  const chart = bb.generate({
-    bindto: "#lineChart",
-    data: {
-      x: "x", // Specifies the x-axis categories (dates)
-      columns: [
-        [
-          "x",
-          "05-02-2025",
-          "06-02-2025",
-          "07-02-2025",
-          "08-02-2025",
-          "09-02-2025",
-        ], // Dates
-        ["Pop-up Promocional", 35, 5, 35, 10, 7], // Data for "Pop-up Promocional"
-        ["Pop-up Suscripción", 0, 0, 5, 8, 4], // Data for "Pop-up Suscripción"
-        ["Pop-up Productos populares", 28, 10, 28, 6, 5], // Data for "Pop-up Productos populares"
-      ],
-      type: line(),
-    },
-    line: {
-      width: 3, // Set the width of the line here
-    },
-    padding: {
-      bottom: 100,
-    },
-    axis: {
-      x: {
-        type: "timeseries",
-        tick: {
-          format: "%d-%m-%Y",
-          text: {
-            inner: {
-              first: true,
-              last: true,
-            },
-          },
-        },
-      },
-    },
-    tooltip: {
-      contents: function (d, defaultTitleFormat, defaultValueFormat, color) {
-        let tooltipHTML = `<div class="bb-tooltip-custom">
-                           <div class="bb-tooltip-title">
-                            Fecha: ${defaultTitleFormat(d[0].x)}
-                          </div>`;
-
-        d.forEach((item) => {
-          tooltipHTML += `<div class="bb-tooltip-item">
-                            <span class="bb-tooltip-marker" style="background-color:${color(item)};"></span>
-                            ${item.name}: ${defaultValueFormat(item.value)}
-                          </div>`;
-        });
-
-        tooltipHTML += `</div>`;
-        return tooltipHTML;
-      },
-    },
-    point: {
-      show: true,
-    },
-    legend: {
-      padding: 20,
-      item: {
-        tile: {
-          width: 16,
-          height: 16,
-        },
-      },
-    },
-  });
-};
-
 export const initDopplerMenuBehavior = function () {
   $("body").on(
     {
